Move `instance_converters.py`.
