+++ metricflow/plan_conversion/to_sql_plan/dataflow_to_subquery.py
@@ -95,7 +95,7 @@ from metricflow.dataflow.nodes.write_to_data_table import WriteToResultDataTable
 from metricflow.dataflow.nodes.write_to_table import WriteToResultTableNode
 from metricflow.dataset.dataset_classes import DataSet
 from metricflow.dataset.sql_dataset import AnnotatedSqlDataSet, SqlDataSet
-from metricflow.plan_conversion.instance_converters import (
+from metricflow.plan_conversion.instance_set_transforms.instance_converters import (
     AddGroupByMetric,
     AddMetadata,
     AddMetrics,
